Finish type parsing
