+++ crates/wgso/tests/cases_invalid/invalid_uniform_param_type/.expected
@@ -4,3 +4,9 @@ error: found buffer with type `i32`, expected uniform type `u32`
 2 | #run test(value=buffer)
   |                 ^^^^^^
   |
+error: found buffer with type `MyStruct`, expected uniform type `MyStruct`
+ --> ./tests/cases_invalid/invalid_uniform_param_type/same_type_name1.wgsl:7:32
+  |
+7 | #run test_same_type_name(param=arg)
+  |                                ^^^
+  |
