+++ crates/wgso/tests/cases_invalid/invalid_uniform_param_type/same_type_name1.wgsl
@@ -1,0 +1,7 @@
+var<storage, read_write> arg: MyStruct;
+
+struct MyStruct {
+    field: u32,
+}
+
+#run test_same_type_name(param=arg)
