+++ crates/wgso/tests/cases_invalid/invalid_uniform_param_type/same_type_name2.wgsl
@@ -1,0 +1,7 @@
+#shader<compute> test_same_type_name
+
+var<uniform> param: MyStruct;
+
+struct MyStruct {
+    field: i32,
+}
